なんか知らんがint64って仕様書に書いてたから変更

+++ model/todo.go
@@ -49,7 +49,7 @@ type (
 
 	// A DeleteTODORequest expresses ...
 	DeleteTODORequest struct {
-		IDs []int `json:"ids"`
+		IDs []int64 `json:"ids"`
 	}
 	// A DeleteTODOResponse expresses ...
 	DeleteTODOResponse struct{}
